Add support for signed metadata
OTA zips are doubly signed: the `payload.bin` file is signed and the
entire zip is also signed. Previously, Custota only relied on the
`payload.bin` signature. This is safe against installing tampered
payloads, but is vulnerable to attacks on the metadata. An attacker
with access to the OTA server could potentially force a downgrade by
editing an old OTA's metadata to appear newer. If the same key is used
to sign OTAs for multiple devices, the attacker could also cause another
device's payload to be installed, potentially bricking the device.

This commit addresses the problem by introducing a new `.csig` file. It
is a CMS signature that contains all of the `ota-property-files` offsets
and sizes from `META-INF/com/android/metadata.pb`, as well as the sha256
digests of the data in those file sections. Since Custota uses
update_engine's streaming mode, the OTA is never downloaded in full and
thus, it's not possible to verify the whole-file signature.

This csig file allows Custota to verify the signature of all metadata
without needing to download the whole OTA. The csig file includes the
digest for `payload_properties.txt`, which contains the digests for
`payload.bin`, forming a strong tie between the metadata and the
payload. If an attacker transplants all metadata from a new OTA into an
old OTA file, the metadata may successfully validate, but the payload
will fail to install due to mismatched digests (as expected).

This is a breaking change. In addition to a new `.csig` file needing to
be generated for each OTA, the format of the update info file changed.
The update info file now only includes the location for downloading the
OTA and csig. The old `metadata_offset` and `metadata_size` fields have
been removed since they're included in the csig info. The file now also
includes a version field to help deal with future changes (if needed).

The csig and "version 2" update info files are generated by a new tool
called custota-tool. It's written in Rust and uses the wonderful
Rust-Crypto collection of libraries for the CMS operations. There seem
to be very few libraries that can deal with CMS, especially when large
files and detached signatures are involved. Other alternatives are the
Bouncy Castle Java library and using openssl's API directly (in C
because I haven't found a single binding for other languages that wrap
the complete CMS API).

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>