Fix crash when UpdaterService is started for the first time
On the initial installation of Custota, the first app launch will set up
the periodic updater job and Android will fire up one run of
UpdaterService almost immediately. UpdaterService is started with
startForegroundService() and because the user hasn't had the chance to
configure a URL yet, the service immediately calls stopService() and
exits. This is not allowed by Android. Even if a service immediately
exits, it must still call startForeground().

This commit adds the startForeground() call whenever UpdaterService
runs, even if there's nothing to do. To avoid a pointless notification
being shown, the FOREGROUND_SERVICE_IMMEDIATE is not used unless an
actual update check is going to be run.

Fixes: #7

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>