Add support for installing from a local directory
This way, folks who don't want to host a server can still take advantage
of Custota's features, like background installation. If they also sync
OTA files to the device using something like Syncthing, then automatic
updates would work too.

The behavior when installing from a local directory is intentionally
identical to what happens when installing from an OTA server. This means
that local installs require the `.json` update info file, the `.csig`
file, and the OTA itself.

There is a small hole that needs to be punched in the SELinux policy to
allow this feature to work. `update_engine` normally is not allowed to
access files on local storage. To work around this, a couple SELinux
rules are added to allow `update_engine` to read the OTA file descriptor
that Custota provided via binder. Notably, `update_engine` is still
blocked from opening new files. In the event that it is exploited, it
only has read access to the single OTA file.

Also, fun fact: `update_engine` is smart. If the user originally started
an update from an OTA server, but cancelled the update halfway through
and switched to local storage (eg. because the network is slow), then
`update_engine` is still able to pick up where it left off.

Closes: #25

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>