custota-selinux: Replace with sepatch-based implementation
The logic in the old custota_selinux executable was split out and ported
to Rust in the sepatch library so that it can be used in other projects.
This commit switches to using sepatch. The new custota-selinux is no
longer unnecessarily bundled into the APK and now lives directly in the
module zip.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>