UpdaterJob: Add support for falling back to active network
It appears on that for some folks running the Android 15 beta, Android's
job scheduler implementation is broken and returns a null Network
instance. This commit adds support for falling back to the currently
active network if the Require Unmetered Network option is turned off.
Otherwise, the only thing we can do is show a notification to tell the
user to turn that option off.

Closes: #68

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>