Show notification for post-reboot cleanup process after OTA update
On virtual A/B devices, after an OTA is installed and the device is
rebooted, the system will merge the new data from the copy-on-write
volumes to the actual partitions. This process takes a few minutes and
most users aren't aware that it is happening.

It also causes some weird behavior with Custota. If the user tried to
check for updates during this process, the notification would say the
updater is initializing for a few minutes and then report that the OTA
was successfully installed (again). Let's fix this by showing what is
actually going on.

This commit also fixes a couple other notification issues:

* The UpdaterThread.ProgressType.INIT state has been removed since it
  races with onStatusUpdate() and sometimes causes the initialization
  message to be shown longer than necessary (until the next progress
  update). It is not necessary anyway since UpdaterService shows the
  initialization message when it hasn't received any state yet.
* Only show the pause, resume, and cancel notification actions during
  the UPDATE, VERIFY, and FINALIZE states because those are the only
  states where they would work anyway.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>