post-fs-data.sh: Combine seapp snippets from all modules to produce full file
This is more reliable and has better compatibility than trying to mount
things ourselves. Any module that modifies plat_seapp_contents and
includes the same command will produce the same output file, so the
order that Magisk/KernelSU loads and mounts modules does not matter.
Currently, MSD and Custota are the only two modules known to modify this
file.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>