Improve workaround for Android 15 null network bug
We'll now try to query the capabilities of the active network in the
fallback path so that the job will still run if the network satisfies
the unmetered network requirement.

Since the fallback path (querying the active network) is no longer
limited compared to the normal path (JobScheduler providing a valid
network), the notification telling users about the Android 15 bug has
been removed.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>