Add support for incremental OTAs
Generating the incremental OTAs is out of scope for Custota, but if
they're already generated some other way, Custota can make use of them.

The vbmeta digest of the current running system is used to select which
incremental OTA to install. This is cryptographically tied to a specific
OS build, so there is no chance that an incremental OTA gets applied to
an incompatible OS. (Unlike eg. if the Android build fingerprint were
used instead.)

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>