Add launcher shortcut to check for updates
This does the same thing as the normal manual check for updates option,
except with fewer steps. There is a small UX issue where the home
screen's shortcuts popup does not close because we launch a UI-less
Activity to schedule the UpdaterJob. This is still better than having
the UI flicker due to a non-UI-less Activity.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>