Clean up care_map.pb and avoid one time crash if it is owned by a different UID
If Custota gets reinstalled or the module disappears during a boot,
Android will likely assign it a new Linux UID when it "reappears".
Overwriting the existing care_map.pb file would then fail during the
next OTA installation attempt until the exception handler deletes the
old file.

With this change, we'll just always delete it first. This also adds some
logic to delete the file after the post-reboot cleanup, when the OTA
installation fails, or when OTA installation is cancelled.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>