Add option to disable pinning network ID
With how WireGuard's kernel mode sets up the network interfaces, if an
app tries to use network pinning, then netd will block the connection
with an error like:

    uid 10168 can't select networks other than 6896.

On Custota's side, we see -EPERM when binding to the TCP socket.

This commit adds a new debug option for disabling forced network
pinning. When the option is disabled, the unmetered network option is
still guaranteed to be used when scheduling the background jobs.
However, once the job is running, it's only respected on a best effort
basis. After network pinning fails once, it will no longer be attempted
for the duration of the job.

With the default settings, if the user encounters this issue, a new
notification will be shown instructing the user to look at the debug
setting.

Fixes: #39

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>