Add support for direct boot
Automatic checks and automatic installs can now run before the user has
unlocked the device following a reboot.

To avoid leaking information (eg. in case the device was stolen), only
update check and success notifications are shown on the lock screen. The
installation progress and failure messages are hidden until the device
is unlocked. This also means an ongoing update cannot be paused or
cancelled while the device is locked without physically interrupting the
internet connection.

If the user uses a local directory or any other SAF provider as the
source, then direct boot support is disabled. Local directories are not
accessible and we can't know whether a SAF provider also supports direct
boot.

The user's preferences will automatically be migrated from
credential-protected storage to device-protected storage. If the app is
downgraded, it will appear as if all settings were reset since the old
preferences file will no longer exist.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>