Fix preferences migration improperly detecting BFU
The migration logic introduced in 5.12 was incorrect and allowed the
migration to run while BFU. This marked the migration as completed, even
though nothing was migrated, causing all settings to be reset.

This commit fixes the broken migration and also ensures that it will
actually run when the user unlocks their device. If the user hasn't
already reconfigured the app after the 5.12 upgrade, then the migration
will be reattempted, even though it is marked as complete.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>