Show determinate progress bar for snapuserd merge
For some reason, update_engine only provides intermediate progress
updates for the snapuserd merge operation if another callback is
registered using IUpdateEngine.cleanupSuccessfulUpdate(). The primary
callback only receives completion statuses. Weirdly, there's also no
binder call to unregister this callback. It can't be removed until the
merge operation completes or the binder client process dies. That is
fine for Custota's use case since UpdaterThread does not exit until the
operation is complete anyway.

This commit also updates the notification logic so that progress
notifications that do not contain action buttons, like the one for merge
operations, are allowed to be shown on the lock screen. This way, the
user can see the progress while in BFU.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>